+++ src/bin/todor/clap_app.rs
@@ -1,6 +1,6 @@
-use clap::{clap_app, ArgMatches};
+use clap::{clap_app, App};
 
-pub fn get_cli_matches<'a>() -> ArgMatches<'a> {
+pub fn build_cli() -> App<'static, 'static> {
     // TODO: add subcommand for just content so it can be piped
     clap_app!(todo_r =>
         (version: env!("CARGO_PKG_VERSION"))
@@ -23,5 +23,4 @@ pub fn get_cli_matches<'a>() -> ArgMatches<'a> {
             (author: "Lavi Blumberg <lavifb@gmail.com>")
         )
     )
-    .get_matches()
 }