+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Fixed -->
 <!-- ### Security -->
 
+## [Unreleased]
+### Added
+- completions for bash, zsh, fish, and powershell
+
+
 ## v0.5.1 (2018-12-19)
 ### Fixed
 - Fixed output when tags are styled to be underlined